Common base class for firedrake transient solvers
